requirements.txt
streamlit
pandas
yfinance
pils
urllib
jinja2